refactor: remove RipaFormContainerMixin, update methods to events, update unit tests
